pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "three-stars"
version = "0.1.0"
description = "Deploy AI-powered web applications to AWS with a single command"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "icoxfog417" },
]
keywords = ["aws", "bedrock", "agentcore", "cloudfront", "cli", "ai", "deployment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "boto3>=1.35",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "moto[s3,cloudfront,iam,sts]>=5.0",
    "ruff>=0.9",
]

[project.scripts]
three-stars = "three_stars.cli:main"

[project.urls]
Homepage = "https://github.com/icoxfog417/three-stars"
Repository = "https://github.com/icoxfog417/three-stars"
Issues = "https://github.com/icoxfog417/three-stars/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/three_stars"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
